+++ tests/testthat/test-plotly-linetype.R
@@ -2,7 +2,7 @@ context("plotly-linetype")
 
 expect_traces <- function(p, n.traces, name){
   stopifnot(is.numeric(n.traces))
-  L <- save_outputs(p, paste0("plotly-", name))
+  L <- save_outputs(p, paste0("plotly-linetype-", name))
   expect_equal(length(L$data), n.traces)
   L
 }
@@ -39,4 +39,26 @@ test_that("Can avoid scaling", {
 test_that("Warn about invalid linetypes", {
   p <- plot_ly(x = 1:2, y = 1:2, linetype = I("DNE"))
   expect_warning(plotly_build(p), "DNE")
+})
+
+test_that("Can specify a scale manually", {
+  pal <- c("1" = "dot", "0" = "dash")
+  p <- plot_ly(mtcars, x = ~mpg, y = ~disp, linetype = ~factor(vs), linetypes = pal)
+  l <- expect_traces(p, 2, "manual")
+  dashes <- lapply(l$data, "[[", "line")
+  expected <- setNames(pal[sapply(l$data, "[[", "name")], NULL)
+  expect_equal(expected, sapply(dashes, "[[", "dash"))
+})
+
+test_that("Trace ordering matches factor levels", {
+  p <- plot_ly(mtcars, x = ~mpg, y = ~disp, linetype = ~factor(vs, levels = c(1, 0))) %>% add_lines()
+  l <- expect_traces(p, 2, "ordering")
+  expect_equal(sapply(l$data, "[[", "name"), c("1", "0"))
+})
+
+test_that("Trace ordering is alphabetical", {
+  lvls <- sort(unique(mpg$class))
+  p <- plot_ly(mpg, x = ~cty, y = ~hwy, linetype = ~class) %>% add_lines()
+  l <- expect_traces(p, length(lvls), "alphabetical")
+  expect_equal(sapply(l$data, "[[", "name"), lvls)
 })