+++ tests/testthat/test-plotly-symbol.R
@@ -2,7 +2,7 @@ context("plotly-symbol")
 
 expect_traces <- function(p, n.traces, name){
   stopifnot(is.numeric(n.traces))
-  L <- save_outputs(p, paste0("plotly-", name))
+  L <- save_outputs(p, paste0("plotly-symbol-", name))
   expect_equal(length(L$data), n.traces)
   L
 }
@@ -12,7 +12,7 @@ test_that("Mapping a variable to symbol works", {
   l <- expect_traces(p, 3, "scatterplot-symbol")
   markers <- lapply(l$data, "[[", "marker")
   syms <- unlist(lapply(markers, "[[", "symbol"))
-  expect_identical(syms, c("circle", "cross", "diamond"))
+  expect_identical(syms, c("circle", "triangle-up", "square"))
 })
 
 test_that("Can set the symbol range.", {
@@ -45,4 +45,26 @@ test_that("Formula resulting in logical vector works", {
   markers <- lapply(l$data, "[[", "marker")
   syms <- unlist(lapply(markers, "[[", "symbol"))
   expect_identical(syms, s)
+})
+
+test_that("Can specify a scale manually", {
+  pal <- c("1" = "cross", "0" = "diamond")
+  p <- plot_ly(mtcars, x = ~mpg, y = ~disp, symbol = ~factor(vs), symbols = pal)
+  l <- expect_traces(p, 2, "symbol-manual")
+  markers <- lapply(l$data, "[[", "marker")
+  expected <- setNames(pal[sapply(l$data, "[[", "name")], NULL)
+  expect_equal(expected, sapply(markers, "[[", "symbol"))
+})
+
+test_that("Trace ordering matches factor levels", {
+  p <- plot_ly(mtcars, x = ~mpg, y = ~disp, symbol = ~factor(vs, levels = c(1, 0)))
+  l <- expect_traces(p, 2, "ordering")
+  expect_equal(sapply(l$data, "[[", "name"), c("1", "0"))
+})
+
+test_that("Trace ordering is alphabetical", {
+  lvls <- sort(unique(mpg$class))
+  p <- plot_ly(mpg, x = ~cty, y = ~hwy, symbol = ~class)
+  l <- expect_traces(p, length(lvls), "alphabetical")
+  expect_equal(sapply(l$data, "[[", "name"), lvls)
 })